[LITO-151] refactor: chat-gpt model, role 고정값으로 변경 [LITO-152] refactor: prompt 엔지니어링 적용
