+++ api/src/main/java/com/swm/lito/api/file/adapter/in/web/FileController.java
@@ -1,4 +1,4 @@
-package com.swm.lito.api.file.adapter.in.presentation;
+package com.swm.lito.api.file.adapter.in.web;
 
 import com.swm.lito.core.common.security.AuthUser;
 import com.swm.lito.core.file.application.port.in.FileUseCase;
