+++ api/src/test/java/com/swm/lito/api/chat/adapter/in/web/ChatControllerTest.java
@@ -1,4 +1,4 @@
-package com.swm.lito.api.chat.adapter.in.presentation;
+package com.swm.lito.api.chat.adapter.in.web;
 
 
 import com.swm.lito.api.chat.adapter.in.request.ChatGPTCompletionRequest;
