+++ core/src/main/java/com/swm/lito/core/chat/application/port/in/request/ChatGPTCompletionRequestDto.java
@@ -1,14 +1,9 @@
 package com.swm.lito.core.chat.application.port.in.request;
 
-import com.theokanning.openai.completion.chat.ChatCompletionRequest;
-import com.theokanning.openai.completion.chat.ChatMessage;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Getter
 @NoArgsConstructor
@@ -22,20 +17,18 @@ public class ChatGPTCompletionRequestDto {
 
     private String message;
 
-    private Integer maxTokens;
-
-    public static ChatCompletionRequest from(ChatGPTCompletionRequestDto request) {
-        return ChatCompletionRequest.builder()
-                .model(request.getModel())
-                .messages(convertChatMessage(request))
-                .maxTokens(request.getMaxTokens())
+    public static ChatGPTCompletionRequestDto from(String message){
+        return ChatGPTCompletionRequestDto.builder()
+                .message(message)
                 .build();
     }
 
-    private static List<ChatMessage> convertChatMessage(ChatGPTCompletionRequestDto request) {
-        List<ChatMessage> chatMessages = new ArrayList<>();
-        chatMessages.add(new ChatMessage("system", "너는 IT 전문가로서 행동하고 대답해야한다."));
-        chatMessages.add(new ChatMessage(request.getRole(), request.getMessage()));
-        return chatMessages;
+    public static ChatGPTCompletionRequestDto from(String model, String role, String message){
+        return ChatGPTCompletionRequestDto.builder()
+                .model(model)
+                .role(role)
+                .message(message)
+                .build();
     }
+
 }
