+++ core/src/main/java/com/swm/lito/core/chat/domain/Chat.java
@@ -5,7 +5,7 @@ import jakarta.persistence.Id;
 import lombok.*;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-@Document(collation = "chats")
+@Document(collection = "chats")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
